refactor: make Server class framework-agnostic by moving express to separate module
- Move createMcpExpressApp and CreateMcpExpressAppOptions to src/server/express.ts
- Remove express import from src/server/index.ts to make Server class usable in browser environments
- Update all imports to use server/express.js instead of server/index.js for createMcpExpressApp
- Update documentation and comments to reference the new import path

This allows Server class to be used without forcing express dependency, making it framework-agnostic and browser-compatible.
